Refactor Part 1
Unified logger
Made logger output prettier
Removed most of the code from cfg.py and seperated into repective files
Update license year
